requirements.txt
python-dotenv==0.20.0
requests==2.22.0
telegram==0.0.1
